Moved part2 and part4 code into seperate .h files
This allows me to create a FileReader class for part 2 and a
DataViewer class for part 4, both with a .run() method that can
be called from the main method.

This isolates both of my questions making it easier to read and
work with
